+++ components/date-time-picker/src/date-time-picker.ts
@@ -10,6 +10,8 @@ export const dateTimePickerModes = [
   'date',
   'datetime',
   'time',
+  'datetimeNoSecond',
+  'timeNoSecond',
 ] as const
 
 export const dateTimePickerProps = buildProps({